✨feat: Implement certificate ID from database
feat: Implement certificate ID from database.

(certificate.tsx): Removes the generation of uuid for the certificate id and instead uses the certificateId passed in the props.
(Preview.tsx): Passes the awardId to the sendInvitationEmail function.
(invite-members.tsx): Deletes the file as this functionality is no longer needed.
(storeEvent.ts): Returns the awardId from the storeEvent function.